requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
httpx>=0.27.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
alembic>=1.13.2
celery>=5.4.0
redis>=5.0.4
python-dotenv>=1.0.1
loguru>=0.7.2
tenacity>=8.3.0
beautifulsoup4>=4.12.3
lxml>=5.2.0
readability-lxml>=0.8.1
playwright>=1.46.0
ruff>=0.5.5
pytest>=8.2.0

Jinja2>=3.1.4
aiofiles>=24.1.0

google-genai>=1.2.0
duckduckgo-search>=6.2.10
jsonschema>=4.23.0

playwright>=1.46.0
beautifulsoup4>=4.12.3
readability-lxml>=0.8.1
lxml>=5.2.2
